feat: Add CDN URL configuration option
This commit adds a new configuration option `cdnUrl` to the `BaseStatic` class in the `index.ts` file. The `cdnUrl` property is of type string and is used to specify the CDN URL for static assets. The `getCdnUrl` method is also added to the class to handle the logic of retrieving the CDN URL.

The `replaceImportPaths` method is modified to prepend the CDN URL to import paths that start with "/static" or "/npm". This ensures that the correct URL is used when importing static assets.

This change improves the flexibility and performance of the static module by allowing the use of a CDN for serving static assets.